update: variant types - moved tmburType and msiType out from mutation burden into common DEVSU-1959
