+++ app/views/ReportView/components/GenomicSummary/index.tsx
@@ -23,6 +23,7 @@ import withLoading, { WithLoadingInjectedProps } from '@/hoc/WithLoading';
 import PatientEdit from '@/components/PatientEdit';
 import useConfirmDialog from '@/hooks/useConfirmDialog';
 import TumourSummaryEdit from '@/components/TumourSummaryEdit';
+import { TumourSummaryType, TmburType, MsiType } from '@/common';
 import VariantChips from './components/VariantChips';
 import VariantCounts from './components/VariantCounts';
 import {
@@ -31,13 +32,12 @@ import {
   MicrobialType,
 } from './types';
 import {
-  MutationBurdenType, ComparatorType, MsiType, TmburType,
+  MutationBurdenType, ComparatorType,
 } from '../MutationBurden/types';
 import MutationSignatureType from '../MutationSignatures/types';
 import { ImmuneType } from '../Immune/types';
 
 import './index.scss';
-import { TumourSummaryType } from '@/common';
 
 const variantCategory = (variant) => {
   // small mutations
