+++ app/views/ReportView/components/MutationBurden/types.d.ts
@@ -24,34 +24,7 @@ type MutationBurdenType = {
   truncatingSnvCount: number | null;
 } & RecordDefaults;
 
-type TmburType = {
-  tumour: string;
-  normal: string;
-  nonNBasesIn1To22AndXAndY: string;
-  totalGenomeSnvs: number;
-  totalGenomeIndels: number;
-  genomeSnvTmb: number;
-  genomeIndelTmb: number;
-  cdsBasesIn1To22AndXAndY: string;
-  cdsSnvs: number,
-  cdsIndels: number;
-  cdsSnvTmb: number;
-  cdsIndelTmb: number;
-  proteinSnvs: number;
-  proteinIndels: number;
-  proteinSnvTmb: number;
-  proteinIndelTmb: number;
-  msiScore: number;
-} & RecordDefaults;
-
-type MsiType = {
-  score: number | null;
-  kbCategory: string | null;
-} & RecordDefaults;
-
 export {
   ComparatorType,
   MutationBurdenType,
-  TmburType,
-  MsiType,
 };
